requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
Werkzeug
Pillow
google-generativeai
